Fix most tests 🏗️
