+++ bootstrap/pentagram/interpret/term.py
@@ -1,28 +1,27 @@
 from __future__ import annotations
 
 from pentagram.machine import MachineCall
-from pentagram.machine import MachineFrame
 from pentagram.machine import MachineFrameStack
-from pentagram.machine import MachineInstructionPointer
 from pentagram.machine import MachineNumber
 from pentagram.machine import MachineValue
-from pentagram.syntax import SyntaxBlock
 from pentagram.syntax import SyntaxComment
+from pentagram.syntax import SyntaxExpression
 from pentagram.syntax import SyntaxIdentifier
 from pentagram.syntax import SyntaxNumber
 from typing import Any
 
 
-def interpret_term(frame_stack: MachineFrameStack) -> None:
-    term = frame_stack.current.term
+def interpret_expression_term(
+    frame_stack: MachineFrameStack,
+    statement: SyntaxExpression,
+) -> None:
+    term = statement.terms[frame_stack.current.term_index]
     if isinstance(term, SyntaxNumber):
         interpret_number_term(frame_stack, term)
     elif isinstance(term, SyntaxIdentifier):
         interpret_identifier_term(frame_stack, term)
     elif isinstance(term, SyntaxComment):
         next_term(frame_stack)
-    elif isinstance(term, SyntaxBlock):
-        interpret_block_term(frame_stack, term)
     else:
         raise AssertionError(term)
 
@@ -50,21 +49,6 @@ def interpret_identifier_term(
         value_or_call(frame_stack)
     else:
         raise AssertionError(value_or_call)
-
-
-def interpret_block_term(
-    frame_stack: MachineFrameStack, block: SyntaxBlock
-) -> None:
-    next_term(frame_stack)
-    frame_stack.push(
-        MachineFrame(
-            instruction_pointer=MachineInstructionPointer(
-                block=block, statement_index=0, term_index=0
-            ),
-            expression_stack=frame_stack.current.expression_stack,
-            environment=frame_stack.current.environment.extend(),
-        )
-    )
 
 
 def next_term(frame_stack: MachineFrameStack) -> None:
